+++ WoWsShipBuilder.Core/Settings/DispersionPlotSettings.cs
@@ -1,0 +1,11 @@
+﻿namespace WoWsShipBuilder.Core.Settings
+{
+    public class DispersionPlotSettings
+    {
+        public int ShotsNumber { get; set; } = 100;
+
+        public double AimingRange { get; set; } = 10.0;
+
+        public bool IsVertical { get; set; } = true;
+    }
+}